Made sakusen build again on windows. Some warning hunting. A comment in the build routine.
git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@814 6f43db7c-3df8-0310-95f7-eae406835aab

+++ libsakusen/server/player.cpp
@@ -109,13 +109,13 @@ void Player::detachClient(Client* client)
 }
 
 uint32 Player::requestEnergy(uint32 amount) {
-  uint32 result = std::min<uint64>(amount, availableEnergy);
+  uint32 result = static_cast<uint32>(std::min<uint64>(amount, availableEnergy));
   availableEnergy -= result;
   return result;
 }
 
 uint32 Player::requestMetal(uint32 amount) {
-  uint32 result = std::min<uint64>(amount, availableMetal);
+  uint32 result = static_cast<uint32>(std::min<uint64>(amount, availableMetal));
   availableMetal -= result;
   return result;
 }
